src_test/tool_test/selenium/19-24/requirements.txt
selenium>=4.20
pytest>=8.0
pytest-html>=4.1
openpyxl>=3.1
mysql-connector-python>=8.4
redis>=5.0
python-dotenv>=1.0
bcrypt>=4.1
requests>=2.31
